Refactor Rule to FilterRule
